+++ src/main/java/romanow/abc/desktop/view/GUIESSWorkSettingBoolean.java
@@ -94,7 +94,7 @@ public class GUIESSWorkSettingBoolean extends GUIElement {
         }
 
     @Override
-    public void putValue(int vv) throws UniException {
+    public void putValue(long vv) throws UniException {
         }
     @Override
     public String setParams(FormContext context, MetaExternalSystem meta, MetaRegister register0, MetaGUIElement element0, I_ModbusGroupDriver plm0, I_GUIEvent back0) {
