+++ src/main/java/romanow/abc/desktop/view2/desktop/DesktopGUIButton.java
@@ -85,7 +85,7 @@ public class DesktopGUIButton extends View2BaseDesktop {
         new Message(200,200,"Команда "+toHex(register.getRegNum())+":="+cmdCode+"$"+ ss,Values.PopupMessageDelay);
         }
     @Override
-    public void putValue(int vv) throws UniException {}
+    public void putValue(long vv) throws UniException {}
 
     @Override
     public String setParams(FormContext2 context, ESS2Architecture meta, Meta2GUI element0, I_GUI2Event onEvent0) {
