+++ src/main/java/romanow/abc/desktop/view2/desktop/DesktopGUILabel.java
@@ -30,5 +30,5 @@ public class DesktopGUILabel extends View2BaseDesktop {
         return null;
         }
     @Override
-    public void putValue(int vv) throws UniException {}
+    public void putValue(long vv) throws UniException {}
 }
