+++ src/main/java/romanow/abc/desktop/view2/desktop/DesktopGUILevelMultiIndicator.java
@@ -97,7 +97,7 @@ public class DesktopGUILevelMultiIndicator extends View2BaseDesktop {
         panel.add(textField3);
         }
     @Override
-    public void putValue(Meta2Register register, int xx, int idx) {
+    public void putValue(Meta2Register register, long xx, int idx) {
         if (register instanceof Meta2SettingRegister)
             limits[idx] = register.doubleWithPower(xx);
         else{
@@ -108,7 +108,7 @@ public class DesktopGUILevelMultiIndicator extends View2BaseDesktop {
             }
         }
     @Override
-    public void putValue(int xx) throws UniException {
+    public void putValue(long xx) throws UniException {
         Meta2DataRegister register = (Meta2DataRegister) getRegister();
         vv[0] = register.doubleWithPower(xx);
         }
